code/requirements.txt
altgraph==0.16.1
apturl==0.5.2
asn1crypto==0.24.0
atomicwrites==1.2.1
attrs==18.2.0
Brlapi==0.6.6
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.0
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
decorator==4.4.0
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
fbs==0.8.3
ffmpeg-python==0.1.17
future==0.17.1
gitignore-parser==0.0.3
gstreamer-player==1.1.2
httplib2==0.9.2
idna==2.8
imageio==2.5.0
imageio-ffmpeg==0.3.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
lxml==4.3.4
macaroonbakery==1.1.3
macholib==1.11
Mako==1.0.7
MarkupSafe==1.0
more-itertools==4.3.0
moviepy==1.0.0
mutagen==1.41.1
netifaces==0.10.4
numpy==1.15.4
oauth==1.0.1
olefile==0.45.1
opencv-contrib-python==4.1.0.25
openshot-qt==2.4.4.dev2
pbr==5.3.1
pdf2image==1.4.2
pefile==2019.4.18
pexpect==4.2.1
Pillow==5.4.1
pkg-resources==0.0.0
pluggy==0.8.0
proglog==0.1.9
protobuf==3.0.0
py==1.7.0
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
PyInstaller==3.4
pymacaroons==0.13.0
PyNaCl==1.1.2
PyQt5==5.12.2
PyQt5-sip==4.19.17
pyRFC3339==1.0
pytest==4.0.1
python-apt==1.6.4
python-dateutil==2.6.1
python-debian==0.1.32
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
pyzmq==16.0.2
reportlab==3.4.0
requests==2.21.0
requests-unixsocket==0.1.5
scipy==1.3.0
scour==0.36
SecretStorage==2.3.1
simplejson==3.13.2
sip==4.19.8
six==1.11.0
sk-video==1.1.10
system-service==0.3
systemd-python==234
testresources==2.0.1
tqdm==4.31.1
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.24.3
usb-creator==0.3.3
wadllib==1.3.2
xkit==0.0.0
zope.interface==4.3.2
